Moved Clock to utils package
